add option to profile the executors and map reduce jobs
