+++ gio/map_reduce.go
@@ -1,13 +1,9 @@
 package gio
 
 import (
-	"context"
 	"flag"
 	"fmt"
-	"log"
 	"os"
-	"strconv"
-	"strings"
 	"sync"
 
 	"github.com/chrislusf/gleam/pb"
@@ -26,6 +22,7 @@ type gleamTaskOption struct {
 	HashCode     uint
 	StepId       int
 	TaskId       int
+	IsProfiling  bool
 }
 
 type gleamRunner struct {
@@ -47,6 +44,7 @@ func init() {
 	flag.UintVar(&taskOption.HashCode, "flow.hashcode", 0, "flow hashcode")
 	flag.IntVar(&taskOption.StepId, "flow.stepId", -1, "flow step id")
 	flag.IntVar(&taskOption.TaskId, "flow.taskId", -1, "flow task id")
+	flag.BoolVar(&taskOption.IsProfiling, "gleam.profiling", false, "profiling all steps")
 }
 
 var (
@@ -93,57 +91,4 @@ func Init() {
 		runner.runMapperReducer()
 		os.Exit(0)
 	}
-}
-
-// Serve starts processing stdin and writes output to stdout
-func (runner *gleamRunner) runMapperReducer() {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	stat.FlowHashCode = uint32(runner.Option.HashCode)
-	stat.Stats = []*pb.InstructionStat{
-		&pb.InstructionStat{
-			StepId: int32(runner.Option.StepId),
-			TaskId: int32(runner.Option.TaskId),
-		},
-	}
-
-	if runner.Option.Mapper != "" {
-		if fn, ok := mappers[runner.Option.Mapper]; ok {
-			if err := runner.processMapper(ctx, fn); err != nil {
-				log.Fatalf("Failed to execute mapper %v: %v", os.Args, err)
-			}
-			return
-		} else {
-			log.Fatalf("Failed to find mapper function for %v", runner.Option.Mapper)
-		}
-	}
-
-	if runner.Option.Reducer != "" {
-		if runner.Option.KeyFields == "" {
-			log.Fatalf("Also expecting values for -gleam.keyFields! Actual arguments: %v", os.Args)
-		}
-		if fn, ok := reducers[runner.Option.Reducer]; ok {
-
-			keyPositions := strings.Split(runner.Option.KeyFields, ",")
-			var keyIndexes []int
-			for _, keyPosition := range keyPositions {
-				keyIndex, keyIndexError := strconv.Atoi(keyPosition)
-				if keyIndexError != nil {
-					log.Fatalf("Failed to parse key index positions %v: %v", runner.Option.KeyFields, keyIndexError)
-				}
-				keyIndexes = append(keyIndexes, keyIndex)
-			}
-
-			if err := runner.processReducer(ctx, fn, keyIndexes); err != nil {
-				log.Fatalf("Failed to execute reducer %v: %v", os.Args, err)
-			}
-
-			return
-		} else {
-			log.Fatalf("Failed to find reducer function for %v", runner.Option.Reducer)
-		}
-	}
-
-	log.Fatalf("Failed to find function to execute. Args: %v", os.Args)
 }
